Add files via upload
There is a string, , of lowercase English letters that is repeated infinitely many times. Given an integer, , find and print the number of letter a's in the first  letters of the infinite string.

For example, if the string s='abcac's= 
′
 abcac 
′
  and n=10n=10 , the substring we consider is abcacabcacabcacabcac, the first 1010 characters of her infinite string. There are 44 occurrences of a in the substring.

Function Description
Complete the repeatedString function in the editor below. It should return an integer representing the number of occurrences of a in the prefix of length nn in the infinitely repeating string.

repeatedString has the following parameter(s):

s: a string to repeat
n: the number of characters to consider
Input Format
The first line contains a single string, ss.The second line contains an integer, nn.

Constraints
1\leq|s|\leq 1001≤∣s∣≤100
1\leq n \leq 10^{12}1≤n≤10 
12
 
Output Format
Print a single integer denoting the number of letter a's in the first nn letters of the infinite string created by repeating infinitely many times.


The substring we consider is , the first  characters of the infinite string. There are  occurrences of a in the substring.

Function Description

Complete the repeatedString function in the editor below.

repeatedString has the following parameter(s):

s: a string to repeat
n: the number of characters to consider
Returns

int: the frequency of a in the substring
Input Format

The first line contains a single string, .
The second line contains an integer, .

Constraints

For  of the test cases, .
Sample Input

Sample Input 0